Fixed small bugs, and improved tests

+++ tests/csv_files/colors.py
@@ -4,6 +4,6 @@ array([('Red', 1.0, 0.0, 0.0),
        ('Orange', 1.0, 0.5, 0.0),
        ('Yellow', 1.0, 1.0, 0.0)],
       dtype=[('Name', 'S6'),
-             ('r', float64),
-             ('g', float64),
-             ('b', float64)])
+             ('r', float),
+             ('g', float),
+             ('b', float)])
